Castor3D: Moved OverlayPass to Render/Overlays folder.
