+++ include/Core/Castor3D/Render/Overlays/OverlaysModule.hpp
@@ -34,6 +34,15 @@ namespace castor3d
 	/**
 	*\~english
 	*\brief
+	*	Overlays render pass.
+	*\~french
+	*\brief
+	*	Passe de rendu des overlays.
+	*/
+	class OverlayPass;
+	/**
+	*\~english
+	*\brief
 	*	The overlays renderer.
 	*\~french
 	*\brief
