+++ source/Core/Castor3D/Miscellaneous/LoadingScreen.cpp
@@ -12,8 +12,8 @@
 #include "Castor3D/Overlay/TextOverlay.hpp"
 #include "Castor3D/Render/RenderDevice.hpp"
 #include "Castor3D/Render/Culling/FrustumCuller.hpp"
+#include "Castor3D/Render/Overlays/OverlayPass.hpp"
 #include "Castor3D/Render/Passes/BackgroundRenderer.hpp"
-#include "Castor3D/Render/Passes/OverlayPass.hpp"
 #include "Castor3D/Render/Passes/ForwardRenderTechniquePass.hpp"
 #include "Castor3D/Scene/Camera.hpp"
 #include "Castor3D/Scene/Scene.hpp"
