+++ source/Core/Castor3D/Render/RenderTarget.cpp
@@ -20,8 +20,8 @@
 #include "Castor3D/Render/Culling/FrustumCuller.hpp"
 #include "Castor3D/Render/EnvironmentMap/EnvironmentMap.hpp"
 #include "Castor3D/Render/Node/SceneRenderNodes.hpp"
+#include "Castor3D/Render/Overlays/OverlayPass.hpp"
 #include "Castor3D/Render/Overlays/OverlayRenderer.hpp"
-#include "Castor3D/Render/Passes/OverlayPass.hpp"
 #include "Castor3D/Render/PostEffect/PostEffect.hpp"
 #include "Castor3D/Render/ToneMapping/ToneMapping.hpp"
 #include "Castor3D/Render/ToTexture/Texture3DTo2D.hpp"
